Connect to DB on platform
